test: reuse base class and override DhisAuthenticationApiTest
we can inherit from the same base class and override what's different
